refactor!: Перемещена логика авторизации в `AuthPage`
